开发“修改角色”部分功能 菜单列表函数未开发

+++ sangeng-framework/src/main/java/com/sangeng/domain/dto/UpdateRoleDto.java
@@ -22,7 +22,7 @@ public class UpdateRoleDto {
     private String roleKey;
     private Integer roleSort;
     private String status;
-    private List<Integer> menuIds;
+    private List<Long> menuIds;
     String remark;
 }
